Don't use Utterances for SRT generation
Utterances could theoretically be used for SRT generation,
but the length of the utterances is often so long (25 words to
a paragraph) that it provides a poor user experience.

It also seems like the Utterances are not meant for SRT generation, but
rather semi-realtime sentiment analysis and as such aren't meant to be
good for closed captions.